requirements.txt
django>=4.1, <5.0
django-stubs >= 1.16.0
django-webpack-loader
django_rest_framework
pyyaml
requests
django-cors-headers
pytest-django
uritemplate
coreapi
pytz
mypy >= 1.1.1
djangorestframework-stubs[compatible-mypy]
drf-yasg